+++ tests/queue-get-next-job.spec.js
@@ -203,45 +203,45 @@ module.exports = function () {
         t.equals(group4[0].status, enums.status.active, 'Returned job is active status')
         t.ok(moment.isDate(group4[0].dateStarted), 'Returned job dateStarted is a date')
 
-        // ---------- Testing dateRetry Values ----------
-        t.comment('queue-get-next-job: dateRetry Values')
+        // ---------- Testing dateEnable Values ----------
+        t.comment('queue-get-next-job: dateEnable Values')
         retryJobs = q.createJob(2).map(j => j)
-        retryJobs[0].dateRetry = moment().add(100, 'seconds').toDate()
-        retryJobs[1].dateRetry = moment().add(-100, 'seconds').toDate()
+        retryJobs[0].dateEnable = moment().add(100, 'seconds').toDate()
+        retryJobs[1].dateEnable = moment().add(-100, 'seconds').toDate()
         return q.addJob(retryJobs)
       }).then((retrySavedJobs) => {
         t.equal(retrySavedJobs.length, 2, 'Jobs saved successfully')
         return queueGetNextJob(q)
       }).then((retryGet) => {
-        t.equal(retryGet.length, 1, 'Only one job available based on dateRetry')
+        t.equal(retryGet.length, 1, 'Only one job available based on dateEnable')
         t.equal(retryGet[0].id, retryJobs[1].id, 'Retry job valid')
 
-        // ---------- Testing dateRetry with retryCount ----------
-        t.comment('queue-get-next-job: dateRetry with retryCount')
+        // ---------- Testing dateEnable with retryCount ----------
+        t.comment('queue-get-next-job: dateEnable with retryCount')
         retryJobs = q.createJob(4).map(j => j)
         retryJobs[0].retryCount = 0
-        retryJobs[0].dateRetry = moment().add(-100, 'seconds').toDate()
+        retryJobs[0].dateEnable = moment().add(-100, 'seconds').toDate()
         retryJobs[1].retryCount = 1
-        retryJobs[1].dateRetry = moment().add(-200, 'seconds').toDate()
+        retryJobs[1].dateEnable = moment().add(-200, 'seconds').toDate()
         retryJobs[2].retryCount = 2
-        retryJobs[2].dateRetry = moment().add(-300, 'seconds').toDate()
+        retryJobs[2].dateEnable = moment().add(-300, 'seconds').toDate()
         retryJobs[3].retryCount = 3
-        retryJobs[3].dateRetry = moment().add(-400, 'seconds').toDate()
+        retryJobs[3].dateEnable = moment().add(-400, 'seconds').toDate()
         return q.addJob(retryJobs)
       }).then((retrySavedJobs) => {
         t.equal(retrySavedJobs.length, 4, 'Jobs saved successfully')
         return queueGetNextJob(q)
       }).then((retryGet2) => {
         retryGet2.sort((a, b) => {
-          if (moment(a.dateRetry).isSameOrBefore(b.dateRetry)) return -1
+          if (moment(a.dateEnable).isSameOrBefore(b.dateEnable)) return -1
           return 1
         })
         t.equal(retryGet2.length, 3, 'Jobs retrieved successfully')
         let ids = retryGet2.map(j => j.id)
-        t.ok(!ids.includes(retryJobs[0].id), 'Retrieved in dateRetry order successfully')
-        t.ok(moment().isBefore(retryGet2[0].dateRetry), 'dateRetry for first job is valid')
-        t.ok(moment(retryGet2[0].dateRetry).isBefore(retryGet2[1].dateRetry), 'dateRetry for second job is valid')
-        t.ok(moment(retryGet2[1].dateRetry).isBefore(retryGet2[2].dateRetry), 'dateRetry for third job is valid')
+        t.ok(!ids.includes(retryJobs[0].id), 'Retrieved in dateEnable order successfully')
+        t.ok(moment().isBefore(retryGet2[0].dateEnable), 'dateEnable for first job is valid')
+        t.ok(moment(retryGet2[0].dateEnable).isBefore(retryGet2[1].dateEnable), 'dateEnable for second job is valid')
+        t.ok(moment(retryGet2[1].dateEnable).isBefore(retryGet2[2].dateEnable), 'dateEnable for third job is valid')
         return queueGetNextJob(q)
       }).then((retryGet3) => {
         t.equal(retryGet3.length, 1, 'Last job retrieved successfully')